refactor access token getter code
